+++ services/web/server/tests/unit/with_dbs/03/resource_usage/test_admin_pricing_plans.py
@@ -11,99 +11,12 @@ from unittest.mock import MagicMock
 import pytest
 from aiohttp.test_utils import TestClient
 from faker import Faker
-from models_library.api_schemas_resource_usage_tracker.pricing_plans import (
-    PricingPlanGet,
-    PricingPlanToServiceGet,
-    PricingUnitGet,
-)
 from models_library.resource_tracker import PricingPlanClassification
 from pytest_mock import MockerFixture
 from pytest_simcore.helpers.assert_checks import assert_status
 from pytest_simcore.helpers.webserver_login import UserInfoDict
 from servicelib.aiohttp import status
 from simcore_service_webserver.db.models import UserRole
-
-
-@pytest.fixture
-def mock_rpc_resource_usage_tracker_service_api(
-    mocker: MockerFixture, faker: Faker
-) -> dict[str, MagicMock]:
-    return {
-        ## Pricing plans
-        "list_pricing_plans": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_plans.list_pricing_plans",
-            autospec=True,
-            return_value=[
-                PricingPlanGet.model_validate(
-                    PricingPlanGet.model_config["json_schema_extra"]["examples"][0],
-                )
-            ],
-        ),
-        "get_pricing_plan": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_plans.get_pricing_plan",
-            autospec=True,
-            return_value=PricingPlanGet.model_validate(
-                PricingPlanGet.model_config["json_schema_extra"]["examples"][0],
-            ),
-        ),
-        "create_pricing_plan": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_plans.create_pricing_plan",
-            autospec=True,
-            return_value=PricingPlanGet.model_validate(
-                PricingPlanGet.model_config["json_schema_extra"]["examples"][0],
-            ),
-        ),
-        "update_pricing_plan": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_plans.update_pricing_plan",
-            autospec=True,
-            return_value=PricingPlanGet.model_validate(
-                PricingPlanGet.model_config["json_schema_extra"]["examples"][0],
-            ),
-        ),
-        ## Pricing units
-        "get_pricing_unit": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_units.get_pricing_unit",
-            autospec=True,
-            return_value=PricingUnitGet.model_validate(
-                PricingUnitGet.model_config["json_schema_extra"]["examples"][0],
-            ),
-        ),
-        "create_pricing_unit": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_units.create_pricing_unit",
-            autospec=True,
-            return_value=PricingUnitGet.model_validate(
-                PricingUnitGet.model_config["json_schema_extra"]["examples"][0],
-            ),
-        ),
-        "update_pricing_unit": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_units.update_pricing_unit",
-            autospec=True,
-            return_value=PricingUnitGet.model_validate(
-                PricingUnitGet.model_config["json_schema_extra"]["examples"][0],
-            ),
-        ),
-        ## Pricing plan to service
-        "list_connected_services_to_pricing_plan_by_pricing_plan": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_plans.list_connected_services_to_pricing_plan_by_pricing_plan",
-            autospec=True,
-            return_value=[
-                PricingPlanToServiceGet.model_validate(
-                    PricingPlanToServiceGet.model_config["json_schema_extra"][
-                        "examples"
-                    ][0],
-                )
-            ],
-        ),
-        "connect_service_to_pricing_plan": mocker.patch(
-            "simcore_service_webserver.resource_usage._pricing_plans_admin_api.pricing_plans.connect_service_to_pricing_plan",
-            autospec=True,
-            return_value=PricingPlanToServiceGet.model_validate(
-                PricingPlanToServiceGet.model_config["json_schema_extra"]["examples"][
-                    0
-                ],
-            ),
-        ),
-    }
 
 
 @pytest.fixture
@@ -137,11 +50,13 @@ async def test_get_admin_pricing_endpoints_user_role_access(
 ):
     ## Pricing plans
 
-    url = client.app.router["list_pricing_plans"].url_for()
+    url = client.app.router["list_pricing_plans_for_admin_user"].url_for()
     resp = await client.get(f"{url}")
     await assert_status(resp, expected)
 
-    url = client.app.router["get_pricing_plan"].url_for(pricing_plan_id="1")
+    url = client.app.router["get_pricing_plan_for_admin_user"].url_for(
+        pricing_plan_id="1"
+    )
     resp = await client.get(f"{url}")
     await assert_status(resp, expected)
